refactor app client
